refactor: move servers to server folder + fix imports
