+++ cli/src/mcp/servers/fsServer.ts
@@ -1,7 +1,7 @@
 import type { DustAPI } from "@dust-tt/client";
 import { DustMcpServerTransport } from "@dust-tt/client";
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
-import { InternalMcpService } from "./internalMcpService.js";
+import { InternalMcpService } from "../types/internalMcpService.js";
 
 /**
  * Filesystem MCP service implementation
